requirements.txt
speedtest-cli
psutil
requests
matplotlib
ping3